Extract a channel-neutral RichContent model from the Slack formatter
Slack reply rendering parsed markdown into structure (paragraphs,
headings, code, tables) and rendered it to Block Kit in one place,
`runtime/slack-block-formatting.ts` — so the structure only ever existed
inside the Slack adapter. Split the two concerns:

- New `messaging/content/rich-content.ts`: a channel-neutral `RichContent`
  tree (paragraph | heading | code | table) plus
  `parseMarkdownToRichContent`. No channel-specific imports.
- `slack-block-formatting.ts` now renders that tree via
  `renderRichContentToBlocks(parseMarkdownToRichContent(text))`.

Behavior-preserving: `textToSlackBlocks` and the splitter helpers keep
their signatures, so `send.ts` and the existing tests are untouched and
stay green. This is the seam other channel adapters (Telegram Rich
Messages, Discord Components v2) will render from, instead of each
re-parsing markdown.

Slice 1 of LUM-2618. Moving the Slack renderer under providers/slack/ is
the next step.
